feat: add query parameter support
Parameters are given as extra arguments after the query number,
in NAME[:TYPE]=VALUE format.

TYPE is one of: text, number, datetime, enum.

datetime values are parsed using iso8601.parse_date().
If zoneinfo is not specified, it's assumed to be UTC.
If zoneinfo is specified and is not UTC,
the value is converted to UTC as Dune expects.

Missing TYPE is autodetected, in the following order of precedence:

- datetime, if VALUE can be parsed using iso8601.parse_date()
- number, if VALUE can be parsed using int() or float()
- enum, if a comma is in VALUE
- text otherwise